requirements.txt
Flask
python-dotenv
PyJWT
mysql-connector-python